third_party/blink/web_tests/fast/events/mutation-during-replace-child-2-expected.txt

Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS target.replaceChild(newChild, oldChild); threw exception HierarchyRequestError: Failed to execute 'replaceChild' on 'Node': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
